# Conflicts:
#	integrationTests/singleShard/block/executingMiniblocksSc_test.go

+++ consensus/interface.go
@@ -9,7 +9,7 @@ import (
 
 // Rounder defines the actions which should be handled by a round implementation
 type Rounder interface {
-	Index() int32
+	Index() int64
 	// UpdateRound updates the index and the time stamp of the round depending of the genesis time and the current time given
 	UpdateRound(time.Time, time.Time)
 	TimeStamp() time.Time
